renamed one of the lambda functions to make development easier.
