+++ clean.sh
@@ -3,6 +3,8 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 cd $SCRIPT_DIR
 
+FUNCTIONS_DIR='src/aws_lambda_python'
+
 rm open-tofu/*.generated.tf
 
 rm -r python/lib
@@ -11,8 +13,8 @@ rm -r layer/create_layer
 
 rm layer/layer_content.zip
 
-rm src/aws_lambda_python/lambda_validator/lambda_validator.zip
+rm "${FUNCTIONS_DIR}"/lambda_validator/lambda_validator.zip
 
-rm src/aws_lambda_python/lambda_controller/lambda_controller.zip
+rm "${FUNCTIONS_DIR}"/lambda_controller/lambda_controller.zip
 
-rm src/aws_lambda_python/lambda_caa_resolver/lambda_caa_resolver.zip
+rm "${FUNCTIONS_DIR}"/lambda_caa_resolver/lambda_caa_resolver.zip
