+++ open-tofu/main.tf.template
@@ -1,4 +1,4 @@
-# Init a defaualt aws provider for the api deployment.
+# Init a default aws provider for the api deployment.
 provider "aws" {
   region                  = "{{api-region}}"
   profile                 = "default"
@@ -56,7 +56,7 @@ resource "aws_lambda_function" "lambda_controller" {
     filename      = "../{{source-path}}/lambda_controller/lambda_controller.zip"
     function_name = "open_mpic_lambda_controller_{{deployment-id}}"
     role          = aws_iam_role.open_mpic_lambda_role.arn
-    handler       = "lambda_function.lambda_handler"
+    handler       = "mpic_orchestrator_lambda_function.lambda_handler"
     source_code_hash = filebase64sha256("../{{source-path}}/lambda_controller/lambda_controller.zip")
     runtime = "python3.11"
     timeout = 60
@@ -72,7 +72,6 @@ resource "aws_lambda_function" "lambda_controller" {
         hash_secret = {{hash-secret}}
       }
     }
-
 }
