+++ src/aws_lambda_python/lambda_controller/mpic_orchestrator_lambda_function.py
